transcription-service/requirements.txt
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.17.0
fsspec==2025.2.0
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
protobuf==5.29.3
pycryptodomex==3.21.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
whisper==1.1.10
yt-dlp==2025.2.11.232920.dev0
